+++ frontend/delightful-flow/src/DelightfulFlow/components/FlowMaterialPanel/hooks/useMaterialSearch.ts
@@ -18,7 +18,7 @@ export default function useMaterialSearch({ tab }:UseMaterialSearchProps) {
 
 	const onSearchChange = useMemoizedFn((e: React.ChangeEvent<HTMLInputElement>) => {
 		setKeyword(e.target.value)
-        // Trigger业务测传入ofupdatedatamethod
+        // Trigger business provided update data method
         switch(tab) {
             case TabObject.Flow:
                 subFlow?.searchListFn?.(e.target.value)