+++ frontend/delightful-flow/src/DelightfulFlow/components/FlowPopup/PopupNode/index.tsx
@@ -165,7 +165,7 @@ export default function PopupNode({ node, showIcon = true, inGroup }: PopupNodeP
 		}
 		if (source) {
 			const newEdges = [
-				// 出边
+				// outgoing edge
 				{
 					id: generateSnowFlake(),
 					source,